requirements.txt
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.9.4
drf-yasg==1.16.0
gunicorn==19.9.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
pkg-resources==0.0.0
psycopg2==2.8.3
pyrsistent==0.15.2
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
ruamel.yaml==0.15.97
six==1.12.0
sqlparse==0.3.0
swagger-spec-validator==2.4.3
uritemplate==3.0.0
urllib3==1.25.3
whitenoise==4.1.2
